refactor: MAT-214 교체 선수 ListItem 컴포넌트 이름 개선
