--- src/components/SubstitutionPlayerList/PlayerListItem/index.ts
@@ -1,1 +1,0 @@
-export * from './PlayerListItem';
