+++ src/components/SubstitutionPlayerList/SubstitutionPlayerListItem/SubstitutionPlayerListItem.css.ts
