+++ src/components/SubstitutionPlayerList/SubstitutionPlayerListItem/index.ts
@@ -1,0 +1,1 @@
+export * from './SubstitutionPlayerListItem';
